fix for Linked List addElement by position

+++ src/com/drivers/ListDriver.java
@@ -8,7 +8,7 @@ public class ListDriver {
 
     public static void main (String[] args) throws LinkedListIndexOutOfBoundsException {
         IList<Integer> l = new LinkedList<>();
-        l.addElement(new Integer(1));
+        l.addElement(new Integer(3));
         l.addElement(new Integer(1));
         l.addElement(new Integer(1));
         l.addElement(new Integer(2), 0);
